requirements.txt
fastapi>=0.70.0
uvicorn>=0.15.0
boto3>=1.24.0
pydantic>=1.9.0
pydantic[email]>=1.9.0
python-multipart>=0.0.5
email-validator>=1.1.3
python-dotenv>=0.20.0
sqlalchemy>=1.4.0